+++ x-pack/plugins/spaces/public/management/create_space/create_space_page.test.tsx
@@ -6,7 +6,6 @@
  */
 
 import type { EuiCheckboxProps } from '@elastic/eui';
-import { EuiButton } from '@elastic/eui';
 import { waitFor } from '@testing-library/react';
 import type { ReactWrapper } from 'enzyme';
 import React from 'react';
@@ -23,7 +22,6 @@ import type { SolutionView, Space } from '../../../common/types/latest';
 import { EventTracker } from '../../analytics';
 import type { SpacesManager } from '../../spaces_manager';
 import { spacesManagerMock } from '../../spaces_manager/mocks';
-import { ConfirmAlterActiveSpaceModal } from '../components/confirm_alter_active_space_modal';
 import { EnabledFeatures } from '../components/enabled_features';
 
 // To be resolved by EUI team.
@@ -153,8 +151,8 @@ describe('ManageSpacePage', () => {
       expect(errors).toEqual([
         'Enter a name.',
         'Enter a URL identifier.',
+        'Select a solution.',
         'Enter initials.',
-        'Select one solution.',
       ]);
 
       expect(spacesManager.createSpace).not.toHaveBeenCalled();
@@ -168,7 +166,7 @@ describe('ManageSpacePage', () => {
 
     {
       const errors = wrapper.find('div.euiFormErrorText').map((node) => node.text());
-      expect(errors).toEqual(['Select one solution.']); // requires solution view to be set
+      expect(errors).toEqual(['Select a solution.']); // requires solution view to be set
     }
 
     updateSpace(wrapper, false, 'oblt');
@@ -274,7 +272,13 @@ describe('ManageSpacePage', () => {
       expect(wrapper.find('input[name="name"]')).toHaveLength(1);
     });
 
-    expect(wrapper.find(EnabledFeatures)).toHaveLength(1);
+    // expect visible features table to exist after setting the Solution View to Classic
+    await waitFor(() => {
+      // switch to classic
+      updateSpace(wrapper, false, 'classic');
+      // expect visible features table to exist again
+      expect(wrapper.find(EnabledFeatures)).toHaveLength(1);
+    });
   });
 
   it('hides feature visibility controls when not allowed', async () => {
@@ -333,9 +337,6 @@ describe('ManageSpacePage', () => {
       await Promise.resolve();
 
       wrapper.update();
-
-      // default for create space: expect visible features table to exist
-      expect(wrapper.find(EnabledFeatures)).toHaveLength(1);
     });
 
     await waitFor(() => {
@@ -350,147 +351,6 @@ describe('ManageSpacePage', () => {
       updateSpace(wrapper, false, 'classic');
       // expect visible features table to exist again
       expect(wrapper.find(EnabledFeatures)).toHaveLength(1);
-    });
-  });
-
-  it('allows a space to be updated', async () => {
-    const spaceToUpdate = {
-      id: 'existing-space',
-      name: 'Existing Space',
-      description: 'hey an existing space',
-      color: '#aabbcc',
-      initials: 'AB',
-      disabledFeatures: [],
-      solution: 'es',
-    };
-
-    const spacesManager = spacesManagerMock.create();
-    spacesManager.getSpace = jest.fn().mockResolvedValue({
-      ...spaceToUpdate,
-    });
-    spacesManager.getActiveSpace = jest.fn().mockResolvedValue(space);
-
-    const onLoadSpace = jest.fn();
-
-    const wrapper = mountWithIntl(
-      <CreateSpacePage
-        spaceId={'existing-space'}
-        spacesManager={spacesManager as unknown as SpacesManager}
-        onLoadSpace={onLoadSpace}
-        getFeatures={featuresStart.getFeatures}
-        notifications={notificationServiceMock.createStartContract()}
-        history={history}
-        capabilities={{
-          navLinks: {},
-          management: {},
-          catalogue: {},
-          spaces: { manage: true },
-        }}
-        eventTracker={eventTracker}
-        allowFeatureVisibility
-        allowSolutionVisibility
-      />
-    );
-
-    await waitFor(() => {
-      wrapper.update();
-      expect(spacesManager.getSpace).toHaveBeenCalledWith('existing-space');
-    });
-
-    expect(onLoadSpace).toHaveBeenCalledWith({
-      ...spaceToUpdate,
-    });
-
-    await Promise.resolve();
-
-    wrapper.update();
-
-    updateSpace(wrapper, true, 'oblt');
-
-    await clickSaveButton(wrapper);
-
-    expect(spacesManager.updateSpace).toHaveBeenCalledWith({
-      id: 'existing-space',
-      name: 'New Space Name',
-      description: 'some description',
-      color: '#AABBCC',
-      initials: 'AB',
-      imageUrl: '',
-      disabledFeatures: ['feature-1'],
-      solution: 'oblt', // solution has been changed
-    });
-
-    expect(reportEvent).toHaveBeenCalledWith('space_solution_changed', {
-      action: 'edit',
-      solution: 'oblt',
-      solution_prev: 'es',
-      space_id: 'existing-space',
-    });
-  });
-
-  it('sets calculated fields for existing spaces', async () => {
-    // The Spaces plugin provides functions to calculate the initials and color of a space if they have not been customized. The new space
-    // management page explicitly sets these fields when a new space is created, but it should also handle existing "legacy" spaces that do
-    // not already have these fields set.
-    const spaceToUpdate = {
-      id: 'existing-space',
-      name: 'Existing Space',
-      description: 'hey an existing space',
-      color: undefined,
-      initials: undefined,
-      imageUrl: undefined,
-      disabledFeatures: [],
-    };
-
-    const spacesManager = spacesManagerMock.create();
-    spacesManager.getSpace = jest.fn().mockResolvedValue({
-      ...spaceToUpdate,
-    });
-    spacesManager.getActiveSpace = jest.fn().mockResolvedValue(space);
-
-    const onLoadSpace = jest.fn();
-
-    const wrapper = mountWithIntl(
-      <CreateSpacePage
-        spaceId={'existing-space'}
-        spacesManager={spacesManager as unknown as SpacesManager}
-        onLoadSpace={onLoadSpace}
-        getFeatures={featuresStart.getFeatures}
-        notifications={notificationServiceMock.createStartContract()}
-        history={history}
-        capabilities={{
-          navLinks: {},
-          management: {},
-          catalogue: {},
-          spaces: { manage: true },
-        }}
-        eventTracker={eventTracker}
-        allowFeatureVisibility
-        allowSolutionVisibility
-      />
-    );
-
-    await waitFor(() => {
-      wrapper.update();
-      expect(spacesManager.getSpace).toHaveBeenCalledWith('existing-space');
-    });
-
-    expect(onLoadSpace).toHaveBeenCalledWith({
-      ...spaceToUpdate,
-    });
-
-    await Promise.resolve();
-
-    wrapper.update();
-
-    // not changing anything, just clicking the "Update space" button
-    await clickSaveButton(wrapper);
-
-    expect(spacesManager.updateSpace).toHaveBeenCalledWith({
-      ...spaceToUpdate,
-      color: '#E7664C',
-      initials: 'ES',
-      imageUrl: '',
     });
   });
 
@@ -528,119 +388,6 @@ describe('ManageSpacePage', () => {
       });
     });
   });
-
-  it('warns when updating features in the active space', async () => {
-    const spacesManager = spacesManagerMock.create();
-    spacesManager.getSpace = jest.fn().mockResolvedValue({
-      id: 'my-space',
-      name: 'Existing Space',
-      description: 'hey an existing space',
-      color: '#aabbcc',
-      initials: 'AB',
-      disabledFeatures: [],
-    });
-    spacesManager.getActiveSpace = jest.fn().mockResolvedValue(space);
-
-    const wrapper = mountWithIntl(
-      <CreateSpacePage
-        spaceId={'my-space'}
-        spacesManager={spacesManager as unknown as SpacesManager}
-        getFeatures={featuresStart.getFeatures}
-        notifications={notificationServiceMock.createStartContract()}
-        history={history}
-        capabilities={{
-          navLinks: {},
-          management: {},
-          catalogue: {},
-          spaces: { manage: true },
-        }}
-        eventTracker={eventTracker}
-        allowFeatureVisibility
-        allowSolutionVisibility
-      />
-    );
-
-    await waitFor(() => {
-      wrapper.update();
-      expect(spacesManager.getSpace).toHaveBeenCalledWith('my-space');
-    });
-
-    await Promise.resolve();
-
-    wrapper.update();
-
-    updateSpace(wrapper);
-
-    await clickSaveButton(wrapper);
-
-    const warningDialog = wrapper.find(ConfirmAlterActiveSpaceModal);
-    expect(warningDialog).toHaveLength(1);
-
-    expect(spacesManager.updateSpace).toHaveBeenCalledTimes(0);
-
-    const confirmButton = warningDialog
-      .find(EuiButton)
-      .find('[data-test-subj="confirmModalConfirmButton"]')
-      .find('button');
-
-    confirmButton.simulate('click');
-
-    await Promise.resolve();
-
-    wrapper.update();
-
-    expect(spacesManager.updateSpace).toHaveBeenCalledTimes(1);
-  });
-
-  it('does not warn when features are left alone in the active space', async () => {
-    const spacesManager = spacesManagerMock.create();
-    spacesManager.getSpace = jest.fn().mockResolvedValue({
-      id: 'my-space',
-      name: 'Existing Space',
-      description: 'hey an existing space',
-      color: '#aabbcc',
-      initials: 'AB',
-      disabledFeatures: [],
-    });
-    spacesManager.getActiveSpace = jest.fn().mockResolvedValue(space);
-
-    const wrapper = mountWithIntl(
-      <CreateSpacePage
-        spaceId={'my-space'}
-        spacesManager={spacesManager as unknown as SpacesManager}
-        getFeatures={featuresStart.getFeatures}
-        notifications={notificationServiceMock.createStartContract()}
-        history={history}
-        capabilities={{
-          navLinks: {},
-          management: {},
-          catalogue: {},
-          spaces: { manage: true },
-        }}
-        eventTracker={eventTracker}
-        allowFeatureVisibility
-        allowSolutionVisibility
-      />
-    );
-
-    await waitFor(() => {
-      wrapper.update();
-      expect(spacesManager.getSpace).toHaveBeenCalledWith('my-space');
-    });
-
-    await Promise.resolve();
-
-    wrapper.update();
-
-    updateSpace(wrapper, false);
-
-    await clickSaveButton(wrapper);
-
-    const warningDialog = wrapper.find(ConfirmAlterActiveSpaceModal);
-    expect(warningDialog).toHaveLength(0);
-
-    expect(spacesManager.updateSpace).toHaveBeenCalledTimes(1);
-  });
 });
 
 function updateSpace(
@@ -676,15 +423,6 @@ function toggleFeature(wrapper: ReactWrapper<any, any>) {
     },
   } = wrapper.find('input#featureCategoryCheckbox_kibana').props() as EuiCheckboxProps;
   onChange({ target: { checked: false } } as any);
-
-  wrapper.update();
-}
-
-async function clickSaveButton(wrapper: ReactWrapper<any, any>) {
-  const saveButton = wrapper.find('button[data-test-subj="save-space-button"]');
-  saveButton.simulate('click');
-
-  await Promise.resolve();
 
   wrapper.update();
 }
